% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_analysis_tools.R
\name{ct_plot}
\alias{ct_plot}
\title{Confusion Matrix Heatmap Plot}
\usage{
ct_plot(class.table, plot.title, conformation)
}
\arguments{
\item{class.table}{A confusion matrix in table or matrix format.}

\item{plot.title}{The title to be displayed on the plot.}

\item{conformation}{A subtitle or caption to be displayed on the plot (e.g., model information).}
}
\value{
A list containing:
  \item{plot}{A heatmap plot visualizing the confusion matrix with counts, percentages, and classification metrics.}
  \item{true_recall}{A data frame of recall values for each class.}
}
\description{
This function generates a heatmap visualization of a confusion matrix using `ggplot2`.
It calculates proportions for True Positives, False Positives, Total Size, Precision,
and Accuracy, and presents them as part of the plot. The plot includes both counts
and percentages of correct/incorrect classifications, along with other summary metrics.
}
\details{
- The function takes a confusion matrix and computes the total row and column sums.
- It calculates proportions for recall, precision, accuracy, and sizes, and uses these metrics to create
  a heatmap. True classifications are marked as 'True' and incorrect ones as 'False'.
- Additional metrics like overall accuracy, class size, and precision are calculated and displayed.
- The heatmap tiles are colored based on the classification outcome, and the alpha level (transparency)
  of the tiles represents the computed proportions for better visual distinction.
- The plot uses `ggplot2` with customized aesthetics, and the plot title and caption can be personalized.
- Class-wise recall values are also computed and returned in the results list.
}
