% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_loader.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Initialization Function}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Character string giving the library directory where the package is installed}

\item{pkgname}{Character string giving the name of the package}
}
\description{
This function sets default options when the package is loaded.
It configures data.frame defaults and glm control parameters
to ensure consistent behavior across package functions.
}
\details{
- Sets `check.names = FALSE` as the default for data.frame creation using the default package
- Sets maximum iterations for glm functions to 100 using stats::glm.control
}
\note{
This function is automatically called when the package is loaded and should not be called directly.
}
