% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_polr.R
\name{fit_polr}
\alias{fit_polr}
\title{Fit a Proportional Odds Logistic Regression (POLR) Model}
\usage{
fit_polr(formula, data)
}
\arguments{
\item{formula}{A formula specifying the model structure (e.g., `class ~ var1 + var2 + ...`).}

\item{data}{A data frame containing the variables in the formula and the response variable.}
}
\value{
A fitted `polr` model object.
}
\description{
This function iteratively fits a proportional odds logistic regression (POLR) model using the
provided formula and dataset. If the initial model fails to converge, the function adjusts the
starting coefficient vector and retries fitting the model until it succeeds.
}
\details{
This function calls \code{fit_polr_model()} to fit a \code{MASS::polr()} model. If the model
fails to converge, it expands the starting coefficient vector and retries fitting the model
in a while loop until successful convergence is achieved.
}
