% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_loader.R
\name{get_loglik_multinom}
\alias{get_loglik_multinom}
\title{Get Log-Likelihood for Multinomial Models}
\usage{
get_loglik_multinom(model)
}
\arguments{
\item{model}{A model object typically of class 'multinom' or 'polr'}
}
\value{
The numeric log-likelihood value
}
\description{
This function safely accesses the unexported logLik.multinom function from the nnet package.
}
\keyword{internal}
