% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_cv_LR.R
\name{k.fold.log.cv}
\alias{k.fold.log.cv}
\title{K-Fold Logistic Cross-Validation}
\usage{
k.fold.log.cv(
  formula,
  data,
  ordinal = T,
  folds = NULL,
  outcome.column = which(colnames(data) == "class"),
  stratify = F,
  sample.vector = floor(round(summary(data$class)/min(summary(data$class))))
)
}
\arguments{
\item{formula}{A formula specifying the model structure.}

\item{data}{A dataframe containing the data to fit the model.}

\item{ordinal}{A boolean indicating whether to fit an ordinal logistic regression (`TRUE`)
or a multinomial logistic regression (`FALSE`). Defaults to `TRUE`.}

\item{folds}{Number of folds for cross-validation. If `NULL`, stratified sampling can be used.}

\item{outcome.column}{The column number of the outcome variable. Defaults to the "class" column.}

\item{stratify}{A boolean indicating whether to stratify the data for sampling. Defaults to `FALSE`.}

\item{sample.vector}{A vector specifying the number of samples to be drawn from each class during stratified sampling.
Defaults to the ratio of the smallest class size.}
}
\value{
A list containing:
        - `acc`: Accuracy of the model across folds.
        - `J.small`: J statistic (Youden's Index) for the smallest class.
        - `ct`: Confusion matrix.
        - `probs`: Dataframe of predicted probabilities and class predictions.
}
\description{
This function performs K-fold cross-validation (CV) for ordinal or multinomial logistic regression
models. It allows stratification of data, calculates performance metrics, and returns accuracy,
the confusion matrix, and other relevant statistics.
}
\details{
- If `folds` is specified, the function performs K-fold cross-validation.
- If `stratify` is `TRUE`, the data is stratified based on `sample.vector` before performing CV.
- Models can either be ordinal (`MASS::polr`) or multinomial (`nnet::multinom`).
- Performance metrics like accuracy and J statistic (for each class) are calculated.
}
