% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_cv_LR.R
\name{k.fold.log.iter}
\alias{k.fold.log.iter}
\title{Iterative K-Fold Logistic Cross-Validation}
\usage{
k.fold.log.iter(
  formula,
  data,
  ordinal = T,
  folds = NULL,
  out.col = which(colnames(data) == "class"),
  stratify = FALSE,
  sample.vector = floor(round(summary(data$class)/min(summary(data$class)))),
  iterations,
  verbose = FALSE
)
}
\arguments{
\item{formula}{A formula specifying the model structure.}

\item{data}{A dataframe containing the data to fit the model.}

\item{ordinal}{A boolean indicating whether to fit an ordinal logistic regression (`TRUE`) or a multinomial
logistic regression (`FALSE`). Defaults to `TRUE`.}

\item{folds}{Number of folds for cross-validation. If `NULL`, stratified sampling can be used.}

\item{out.col}{The column number of the outcome variable. Defaults to the "class" column.}

\item{stratify}{A boolean indicating whether to stratify the data for sampling. Defaults to `FALSE`.}

\item{sample.vector}{A vector specifying the number of samples to be drawn from each class during stratified sampling.
Defaults to the ratio of the smallest class size.}

\item{iterations}{Number of iterations to repeat the cross-validation. Each iteration results in a different random split.}

\item{verbose}{A boolean to control whether to print the classification tables and accuracy results.
Defaults to `FALSE`.}
}
\value{
A list containing:
        - `over.all.accuracy`: Overall average accuracy across all iterations.
        - `best`: The best accuracy obtained across iterations.
        - `worst`: The worst accuracy obtained across iterations.
        - `cts`: A formatted table showing the best and worst confusion matrices.
}
\description{
This function performs multiple iterations of K-fold cross-validation (CV) for ordinal or multinomial
logistic regression models, allowing for repeated assessment of accuracy across multiple random splits.
The results include overall accuracy, best/worst iteration accuracy, and confusion tables for both best
and worst iterations.
}
\details{
- This function repeatedly calls the `k.fold.log.cv` function to perform cross-validation for the specified
  number of iterations.
- The function computes and prints the average accuracy, best and worst accuracies, and corresponding
  confusion matrices if `verbose` is set to `TRUE`.
}
