% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simi_sampler.R
\name{simi.sampler}
\alias{simi.sampler}
\title{Similarity-Based Sampling Function}
\usage{
simi.sampler(
  data,
  class,
  compare.with = 0,
  plot = FALSE,
  sample.size = min(summary(as.factor(data$class)))
)
}
\arguments{
\item{data}{A data frame containing features and class labels. The class label column should be named 'class'.}

\item{class}{The class from which to sample observations based on similarity.}

\item{compare.with}{The index of the class similarity to compare (default is 0, which refers to the "same class" similarity).}

\item{plot}{Logical. If `TRUE`, plots of similarity before and after group truncation will be displayed. Default is `FALSE`.}

\item{sample.size}{The number of observations to sample from the specified class. Default is the size of the smallest class in the data.}
}
\value{
A vector of indices corresponding to the rows in the dataset that have been sampled.
}
\description{
This function samples a subset of observations from a dataset based on their similarity
to class-specific centroids (average class feature vectors). It calculates the similarity
of each observation to its own class as well as to other classes and returns the indices of
the sampled subset. It can optionally plot the similarity before and after sampling.
}
\details{
- **Data Scaling**:
  - The feature columns are scaled to have mean zero and unit variance before similarity calculations.
  - The columns `flag` and `tag` (if present) are excluded from the analysis.
- **Similarity Calculation**:
  - For each class, a centroid vector is computed by averaging the scaled feature vectors of observations in that class.
  - The similarity between each observation and its class's centroid vector is calculated using the cosine similarity formula.
- **Sampling**:
  - The function samples observations from the specified class by evenly selecting points across the similarity range.
  - The sample size is specified by `sample.size`, and the sampled points are those closest to evenly spaced intervals within the range of similarities.
- **Plotting**:
  - If `plot = TRUE`, the function will generate two plots using `ggplot2`:
    - **Before Truncation**: A plot showing the similarity of each instance to its class centroid before sampling.
    - **After Truncation**: A plot showing the similarity of the sampled instances after truncation.
  - Both plots will use colors to represent different classes.
}
