% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_model.R
\name{sub_model_log}
\alias{sub_model_log}
\title{Submodel Logistic Regression (Ordinal/Multinomial)}
\usage{
sub_model_log(
  data,
  out.col = which(colnames(data) == "class"),
  min = 1,
  max = floor(dim(data)[1]/5),
  ordinal = T
)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{out.col}{The column number of the outcome variable. Defaults to the "class" column.}

\item{min}{The minimum number of predictor variables to include in combinations. Defaults to 1.}

\item{max}{The maximum number of predictor variables to include in combinations. Defaults to one-fifth of
the number of rows in the dataset.}

\item{ordinal}{A boolean indicating whether to fit an ordinal logistic regression (`TRUE`) or a multinomial
logistic regression (`FALSE`). Defaults to `TRUE`.}
}
\value{
A dataframe containing the top 10 models sorted by McFadden's R-squared value. The output includes:
        - `formula`: The model formula.
        - `McFadden R2`: McFadden's R-squared value, indicating model fit.
}
\description{
This function performs logistic regression by generating and fitting various combinations of predictor
variables. It can handle both ordinal logistic regression (via `MASS::polr`) and multinomial logistic
regression (via `nnet::multinom`). The models are evaluated using McFadden's R-squared value to rank
model performance.
}
\details{
- This function generates combinations of predictor variables and fits logistic regression models for
  each combination.
- McFadden's R-squared value is used to evaluate model fit, with higher values indicating better fit.
- For ordinal regression, the `MASS::polr` function is used, while multinomial regression uses
  `nnet::multinom`.
- The function handles special characters in variable names by wrapping them in backticks.
}
